+++ ng/nggui.h
@@ -8,7 +8,12 @@
 #include "nggraphics.h"
 
 namespace ng {
-	
+
+	enum EnumShape {
+		ShapeRect = 0,
+		ShapeEllipse = 1
+	};
+
 	enum EnumCanvasMode {
 		CanvasRoot = 0, // root canvas. no translation, no scale.
 		CanvasRect = 1, // translate relative to rect.
@@ -16,72 +21,36 @@ namespace ng {
 		CanvasScale = 3 // scale relative to rect.
 	};
 	
-	/*
-	// Rect
-		int contains (int x, int y) const;
-		int overlaps (const Rect*) const;
-		
-		// move rect by vec, consuming vec in the process (vec -> 0). 2d.
-		void moveby (Vec* const);
-		
-		// move rect to (x, y), and reduce vec to the remaining motion. 2d.
-		// assumes (x, y) is in the direction of vec.
-		void moveto (Vec* const, int x, int y);
-		
-		// return true/edge/false if rect a collided with stationary rect b, and
-		// move rect a by the part of its vec v that gets it to collide, and
-		// reduce vec v to the remaining motion.
-		// edge means they will touch and not move any further.
-		// assumes rects are not overlapping.
-		int collide (Vec* const v, const Rect* b);
-	*/
-	
-	/*
-	class Circle {
-	public:
-		int x;
-		int y;
-		// int z; // todo
-		int r;
-		
-		void init (int x, int y, int r);
-		
-		// return true/edge/false if circle contains (x, y).
-		int contains (int x, int y) const;
-		
-		// return true/edge/false if circles overlap/touch.
-		int overlaps (const Circle*) const;
-		
-		// return true/edge/false if circle a collided with stationary circle b, and
-		// move circle a by the part of its vec v that gets it to collide, and
-		// reduce vec v to the remaining motion.
-		// edge means they will touch and not move any further.
-		// assumes circles are not overlapping.
-		int collide (Vec* const v, const Circle* b);
-	};
-	*/
-	
-	/*
 	class Hitbox {
 	public:
-		int shape; // EnumShape {Square, Circle}
 		// center point
 		int x;
 		int y;
 		// int z; // todo
-		// axis/radius from center. a==b means Square/Circle. a!=b means Rect/Ellipse.
+		// radius from center along each axis
 		int a;
 		int b;
 		// int c;
+		int shape;
 		
-		void init (int shape, int x, int y, int r); // Sets symmetry to Radius
-		void init (int shape, int x, int y, int a, int b); // Sets symmetry to Axis
+		// default shape is rect. change later if you want.
+		void init (int x, int y, int a, int b);
+		void init (const Rect*);
+		void init (const Rect*, const Grid*, int column, int row);
+		
+		int major () const; // longest radius
+		int minor () const; // shortest radius
 		
 		// get a graphics rect for drawing this hitbox.
 		void get_rect (Rect* const) const;
 		
 		void scale (const Scale*);
 		void scale_inv (const Scale*);
+		
+		// Generic Hitbox:
+		// If shapes are the same, it will use the correct collider.
+		// If shapes are different, it uses the collider for the larger shape.
+		int get_shape (const Hitbox* h) const;
 		
 		// true if this contains (x, y).
 		bool contains (int x, int y) const;
@@ -90,19 +59,34 @@ namespace ng {
 		bool overlaps (const Hitbox* h) const;
 		
 		// true if this will collide with h as a result of movement by v.
-		// assumes this is same shape as h.
 		bool collides (const Vec* v, const Hitbox* h) const;
 		
 		// collide this with h, and reduce v to remaining motion.
-		// assumes this collides with, and is the same shape as, h.
-		void collide (Vec* const v, const Hitbox* h);
+		// true if this collides with h.
+		bool collide (Vec* const v, const Hitbox* h);
 		
-		// true if the ray (this.x, this.y, in the direction of v) intersects h.
+		// true if the point vector (this.x, this.y, v.x, v.y) intersects h.
 		bool intersects (const Vec* v, const Hitbox* h) const;
 		
 		// intersect this with h (as defined above), and reduce v to remaining motion.
-		// assumes this intersects h.
-		void intersect (Vec* const v, const Hitbox* h);
+		// true if this intersects h.
+		bool intersect (Vec* const v, const Hitbox* h);
+		
+		// Rect Hitbox:
+		bool rect_contains (int x, int y) const;
+		bool rect_overlaps (const Hitbox* h) const;
+		bool rect_collides (const Vec* v, const Hitbox* h) const;
+		bool rect_collide (Vec* const v, const Hitbox* h);
+		bool rect_intersects (const Vec* v, const Hitbox* h) const;
+		bool rect_intersect (Vec* const v, const Hitbox* h);
+		
+		// Ellipse Hitbox: (or Circle if a==b) TODO
+		bool ellipse_contains (int x, int y) const;
+		bool ellipse_overlaps (const Hitbox* h) const;
+		bool ellipse_collides (const Vec* v, const Hitbox* h) const;
+		bool ellipse_collide (Vec* const v, const Hitbox* h);
+		bool ellipse_intersects (const Vec* v, const Hitbox* h) const;
+		bool ellipse_intersect (Vec* const v, const Hitbox* h);
 		
 		// move this by v, and reduce v to 0.
 		void moveby (Vec* const v);
@@ -110,7 +94,6 @@ namespace ng {
 		// move this to (x, y), and reduce v to remaining motion.
 		void moveto (Vec* const v, int x, int y);
 	};
-	*/
 	
 	class Canvas {
 	public: